#220 Move command executor into a separate namespace
