+++ src/MultiplayerMod.Test/Multiplayer/MultiplayerTools.cs
@@ -4,6 +4,7 @@ using MultiplayerMod.Core.Dependency;
 using MultiplayerMod.Core.Scheduling;
 using MultiplayerMod.Multiplayer;
 using MultiplayerMod.Multiplayer.Configuration;
+using MultiplayerMod.Multiplayer.CoreOperations.CommandExecution;
 using MultiplayerMod.Multiplayer.Objects;
 using MultiplayerMod.Multiplayer.Players;
 using MultiplayerMod.Multiplayer.World;
