+++ src/MultiplayerMod/Multiplayer/Configuration/MultiplayerCoordinator.cs
@@ -6,6 +6,7 @@ using MultiplayerMod.Core.Unity;
 using MultiplayerMod.Game;
 using MultiplayerMod.Multiplayer.Commands;
 using MultiplayerMod.Multiplayer.Components;
+using MultiplayerMod.Multiplayer.CoreOperations.CommandExecution;
 using MultiplayerMod.Multiplayer.CoreOperations.Events;
 using MultiplayerMod.Multiplayer.Players.Events;
 using MultiplayerMod.Multiplayer.UI;
