+++ src/MultiplayerMod/Multiplayer/CoreOperations/CommandExecution/CommandConfigurationException.cs
@@ -1,6 +1,6 @@
 ﻿using MultiplayerMod.Exceptions;
 
-namespace MultiplayerMod.Multiplayer.Configuration;
+namespace MultiplayerMod.Multiplayer.CoreOperations.CommandExecution;
 
 public class CommandConfigurationException : MultiplayerException {
     public CommandConfigurationException(string message) : base(message) { }
