+++ src/MultiplayerMod/Multiplayer/MultiplayerLoader.cs
@@ -4,6 +4,7 @@ using MultiplayerMod.ModRuntime;
 using MultiplayerMod.ModRuntime.Loader;
 using MultiplayerMod.Multiplayer.Configuration;
 using MultiplayerMod.Multiplayer.CoreOperations;
+using MultiplayerMod.Multiplayer.CoreOperations.CommandExecution;
 using MultiplayerMod.Multiplayer.Objects;
 using MultiplayerMod.Multiplayer.World;
